requirements.txt
opencv-contrib-python==4.5.5.62
ghostscript==0.7
camelot-py
numpy==1.22.0
pandas==1.4.1
